src/procstat/pyproject.toml
[project]
name = "procstat"
version = "0.6.8"
description = ""
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [
    { name = "SD Traders", email = "torrino07@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

dependencies = [
    "psutil>=7.0.0",
    "pyzmq>=27.0.0"
]
